Refactor method from Event column component to his own file
